LayoutTests/platform/ios-sim-deprecated/fast/events/continuous-platform-wheelevent-in-scrolling-div-expected.txt

FAIL div.scrollTop should be 200. Was 0.
FAIL div.scrollLeft should be 100. Was 0.
